+++ public/pages/Detectors/components/WazuhUpdateBasicDetails/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Copyright Wazuh Inc.
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+export { WazuhUpdateDetectorBasicDetails } from './WazuhUpdateBasicDetails';